Refactor Decimal to Integer type in watches/cards tables

+++ data/migrations/20201019160707_watches.js
@@ -7,7 +7,7 @@ exports.up = function(knex) {
         .notNullable()
       
       watches
-        .decimal('price')
+        .integer('price')
         .notNullable()
 
       watches
