+++ data/migrations/20201019160715_cards.js
@@ -7,7 +7,7 @@ exports.up = function(knex) {
         .notNullable()
       
       cards
-        .decimal('price')
+        .integer('price')
         .notNullable()
 
       cards
